+++ KinaUnaProgenyApi/Services/KanbanServices/KanbanItemsService.cs
@@ -86,17 +86,27 @@ namespace KinaUnaProgenyApi.Services.KanbanServices
         /// database before attempting to delete it. If the item is not found, no changes are made to the
         /// database.</remarks>
         /// <param name="kanbanItem">The Kanban item to delete. The item must have a valid <see cref="KanbanItem.KanbanItemId"/>.</param>
+        /// <param name="hardDelete">If set to <see langword="true"/>, the Kanban item is permanently removed from the database.</param>
         /// <returns>The deleted Kanban item if it was successfully removed; otherwise, <see langword="null"/> if the item does
         /// not exist in the database. Does not include the associated TodoItem.</returns>
-        public async Task<KanbanItem> DeleteKanbanItem(KanbanItem kanbanItem)
+        public async Task<KanbanItem> DeleteKanbanItem(KanbanItem kanbanItem, bool hardDelete = false)
         {
             KanbanItem existingKanbanItem = await progenyDbContext.KanbanItemsDb.SingleOrDefaultAsync(k => k.KanbanItemId == kanbanItem.KanbanItemId);
             if (existingKanbanItem == null)
             {
                 return null;
             }
-            
-            progenyDbContext.KanbanItemsDb.Remove(existingKanbanItem);
+
+            if (hardDelete)
+            {
+                progenyDbContext.KanbanItemsDb.Remove(existingKanbanItem);
+            }
+            else
+            {
+                existingKanbanItem.IsDeleted = true;
+                progenyDbContext.KanbanItemsDb.Update(existingKanbanItem);
+            }
+
             await progenyDbContext.SaveChangesAsync();
 
             return existingKanbanItem;
@@ -109,14 +119,20 @@ namespace KinaUnaProgenyApi.Services.KanbanServices
         /// populates  their associated to-do item details by retrieving them individually. The returned list will be 
         /// empty if no items are associated with the specified board. This method does not validate if a user has access to the data.</remarks>
         /// <param name="kanbanBoardId">The unique identifier of the Kanban board for which to retrieve the items.</param>
+        /// <param name="includeDeleted">If set to <see langword="true"/>, items marked as deleted will be included in the results.</param>
         /// <returns>A task that represents the asynchronous operation. The task result contains a list of  <see
         /// cref="KanbanItem"/> objects associated with the specified Kanban board. Each item includes  its
         /// corresponding to-do item details.</returns>
-        public async Task<List<KanbanItem>> GetKanbanItemsForBoard(int kanbanBoardId)
+        public async Task<List<KanbanItem>> GetKanbanItemsForBoard(int kanbanBoardId, bool includeDeleted = false)
         {
             List<KanbanItem> kanbanItems = await progenyDbContext.KanbanItemsDb.AsNoTracking().Where(ki => ki.KanbanBoardId == kanbanBoardId).ToListAsync();
             foreach (KanbanItem kanbanItem in kanbanItems)
             {
+                if (!includeDeleted && kanbanItem.IsDeleted)
+                {
+                    continue;
+                }
+
                 kanbanItem.TodoItem = await todosService.GetTodoItem(kanbanItem.TodoItemId);
             }
